EKA-102 | Processing composition if the bundle is document. Introucing CCRDocument as parent component
