+++ src/components/HealthInfoContainer/HipHealthInfoContainer.js
@@ -3,6 +3,7 @@ import Typography from "@material-ui/core/Typography";
 import DiagnosticReportTable from "../../components/DiagnosticReport/DiagnosticReportTable";
 import ObservationTable from "../../components/ObservationTable/ObservationTable";
 import HipHealthInfoContainerStyles from "./HipHealthInfoContainer.style";
+import CCRDocument from "../../components/Composition/CCRDocument";
 
 const HipHealthInfoContainer = ({ consentReqId, hipName, data }) => {
   return (
@@ -11,6 +12,7 @@ const HipHealthInfoContainer = ({ consentReqId, hipName, data }) => {
         <Typography className="header" gutterBottom variant="h5" component="h2">
           {hipName}
         </Typography>
+        <CCRDocument consentReqId={consentReqId} data={data} />
         <ObservationTable data={data} />
         <DiagnosticReportTable consentReqId={consentReqId} data={data} />
       </div>
